[refac] change package structure
