--- src/main/java/org/hankki/hankkiserver/domain/UserInfo.java
@@ -1,41 +1,0 @@
-package org.hankki.hankkiserver.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class UserInfo {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_info_id")
-    private Long id;
-    @OneToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL, orphanRemoval = true)
-    @JoinColumn(name = "user_id")
-    private User user;
-    private String refreshToken;
-    private String nickname;
-    private String profileImageUrl;
-
-    public static UserInfo createMemberInfo(
-            final User user,
-            final String refreshToken) {
-        return UserInfo.builder()
-                .user(user)
-                .refreshToken(refreshToken)
-                .nickname(user.getName())
-                .build();
-    }
-
-    public void updateRefreshToken(final String refreshToken) {
-        this.refreshToken = refreshToken;
-    }
-
-    public void updateNickname(final String nickname) {
-        this.nickname = nickname;
-    }
-}
