+++ src/main/java/org/hankki/hankkiserver/domain/user/model/Platform.java
@@ -2,13 +2,24 @@ package org.hankki.hankkiserver.domain.user.model;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import org.hankki.hankkiserver.common.dto.ErrorMessage;
+import org.hankki.hankkiserver.exception.InvalidValueException;
+
+import java.util.Arrays;
 
 @Getter
 @RequiredArgsConstructor
 public enum Platform {
 
-    KAKAO("KAKAO"), APPLE("APPLE");
+    KAKAO("KAKAO"),
+    APPLE("APPLE");
 
-    private final String value;
+    private final String loginPlatform;
 
+    public static Platform getEnumPlatformFromStringPlatform(String loginPlatform) {
+        return Arrays.stream(values())
+                .filter(platform -> platform.loginPlatform.equals(loginPlatform))
+                .findFirst()
+                .orElseThrow(() -> new InvalidValueException(ErrorMessage.INVALID_PLATFORM_TYPE));
+    }
 }
