+++ src/main/java/org/hankki/hankkiserver/domain/user/model/User.java
@@ -1,41 +1,28 @@
 package org.hankki.hankkiserver.domain.user.model;
 
 import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hankki.hankkiserver.domain.common.BaseTimeEntity;
+
+import lombok.*;
+import org.hankki.hankkiserver.domain.BaseTimeEntity;
 
 import java.time.LocalDateTime;
 
+import static org.hankki.hankkiserver.domain.user.model.MemberStatus.ACTIVE;
+import static org.hankki.hankkiserver.domain.user.model.MemberStatus.INACTIVE;
+import static org.hankki.hankkiserver.domain.user.model.UserRole.USER;
+
 @Entity
 @Getter
-@Table(name = "users")
+@Builder(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Table(name = "users")
 public class User extends BaseTimeEntity {
 
     @Id
-    @Column(name = "user_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "user_id")
     private Long id;
-
-    @OneToOne
-    @JoinColumn(name = "user_info_id")
-    private UserInfo userInfo;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Platform platform;
-
-    @Enumerated(EnumType.STRING)
-    private UserRole UserRole;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private MemberStatus status;
-
-    @Column(nullable = false)
-    private String serialId;
 
     @Column(nullable = false)
     private String name;
@@ -43,6 +30,39 @@ public class User extends BaseTimeEntity {
     @Column(nullable = false)
     private String email;
 
+    @Column(nullable = false)
+    private String serialId;
+
+    @Enumerated(EnumType.STRING)
+    private Platform platform;
+
+    @Enumerated(EnumType.STRING)
+    private UserRole userRole;
+
     private LocalDateTime deletedAt;
 
+    @Enumerated(EnumType.STRING)
+    private MemberStatus memberStatus;
+
+    public static User createUser(
+            final String name,
+            final String email,
+            final String serialId,
+            final Platform platform) {
+        return User.builder()
+                .name(name)
+                .email(email)
+                .serialId(serialId)
+                .platform(platform)
+                .userRole(USER)
+                .memberStatus(ACTIVE)
+                .build();
+    }
+
+    public void softDelete(MemberStatus memberStatus) {
+        this.memberStatus = memberStatus;
+        if (INACTIVE == memberStatus) {
+            this.deletedAt = LocalDateTime.now();
+        }
+    }
 }
